apps/forge-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-engine"
version = "1.0.0"
description = "FORGE Engine - Viral clip processing backend"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "FORGE/LAB Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.23",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
ml = [
    "faster-whisper>=0.10.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.1.0",
    "librosa>=0.10.1",
    "scipy>=1.11.0",
    "scenedetect[opencv]>=0.6.2",
    "watchfiles>=0.21.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "mypy>=1.7.0",
]
whisperx = [
    "whisperx>=3.1.0",
]

[project.scripts]
forge-engine = "forge_engine.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["forge_engine"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

